Extract Cookie class
